app/server/requirements.txt
asgiref==3.9.1
cryptography==41.0.4
dj-database-url==2.1.0
Django==4.2.17
django-cors-headers==4.8.0
django-environ==0.12.0
django-password-validators==1.7.3
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
shapely==2.0.6
pymysql==1.1.1
python-Levenshtein==0.25.1
pillow==10.1.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
openpyxl==3.1.2
gunicorn==21.2.0
whitenoise==6.6.0
